api/requirements.txt
fastapi>=0.100.0,<0.111.0
uvicorn[standard]>=0.20.0,<0.29.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
boto3>=1.28.0,<1.35.0
PyMuPDF>=1.23.0,<1.25.0
python-docx>=1.0.0,<1.2.0
spacy>=3.5.0,<3.8.0
beautifulsoup4>=4.11.0,<4.13.0
